requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3

google-genai==1.70.0

pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.0.0
PyMuPDF==1.24.14

python-dotenv==1.0.1
